Add a party color for each candidate and fix bug related to totalling votes.
Add a partyColor property to the factory object and pass it as a
parameter to call when creating an object from the factory object. Add
corresponding colors to each candidate.

Fix bug in the function that totalled the votes, but left them called
"total", which was only declared locally instead of totalVotes which
could be called as a part of the object. This fixed the "winner" code,
as the functions were === to each other (which is why console.log kept
returning "Draw"), but the actual total number of votes differed (which
had not been linked in code to the politician object via totalVotes
parameter) should not have been equal.